Code Studio question to find duplicate items.
Approach 1: Sort all elements and comapare i and i+1 element. If i==i+1 then return ar[i].
Time complexity O(n2)

Approach 2: Use counting sort Again time complexity O(n2)

Approach 3: XOR

a ^ a = 0
a ^ 0 = a

Step1: in ans store XOR of all integers.
Step2: from 1 to N-1 again do XOR of ans and i

we have given array will be from 1 to N-1 so we are doing  XOR of all elements and a^a^a=a